백승현 : 선형보간으로 Draw가 정밀하게 그려지도록 조치함

+++ GameEngineContents/PlayUI.cpp
@@ -109,7 +109,7 @@ void PlayUI::DebugUIOn()
 
 void PlayUI::UpdateSpeed()
 {
-	float Speed = Speed_;
+	float Speed = static_cast<float>(Speed_);
 	if (nullptr != Player::MainPlayer)
 	{
 		Speed = Player::MainPlayer->GetForwardSpeed();
